+++ micromark-extension-cite/docs/cite-syntax-alt.png
