--- old-micromark-extension-cite/package.json
@@ -1,57 +1,0 @@
-{
-	"name": "@benrbray/micromark-extension-cite",
-	"version": "1.0.0",
-	"license": "MIT",
-	"private": false,
-	"description": "Micromark syntax extension for pandoc-style citations.",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/benrbray/remark-cite.git",
-		"directory": "micromark-extension-cite"
-	},
-	"keywords": [
-		"markdown",
-		"pandoc",
-		"citations",
-		"bibliography",
-		"remark",
-		"remark-plugin",
-		"micromark",
-		"micromark-plugin",
-		"micromark-extension"
-	],
-	"author": {
-		"name": "Benjamin R. Bray",
-		"url": "https://benrbray.com/"
-	},
-	"main": "dist/index.cjs.js",
-	"module": "dist/index.esm.js",
-	"browser": "dist/index.umd.js",
-	"files": [
-		"dist"
-	],
-	"scripts": {
-		"build": "rollup -c",
-		"clean": "npx del dist",
-		"prepare": "npm run clean; npm run build",
-		"pretest": "npm run clean; npm run build",
-		"test": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha -r ts-node/register 'test/**/*.ts'"
-	},
-	"dependencies": {
-		"micromark": "^2.11.4"
-	},
-	"devDependencies": {
-		"@babel/plugin-transform-runtime": "^7.13.15",
-		"@rollup/plugin-babel": "^5.3.0",
-		"@rollup/plugin-commonjs": "^18.1.0",
-		"@types/lodash": "^4.14.168",
-		"@types/mocha": "^8.2.2",
-		"@wessberg/rollup-plugin-ts": "^1.3.14",
-		"del-cli": "^3.0.1",
-		"lodash": "^4.17.21",
-		"mocha": "^8.3.2",
-		"rollup": "^2.47.0",
-		"ts-node": "^9.1.1",
-		"unified": "^9.2.1"
-	}
-}
